- Reorganised code folder to src. - Moved Slug.php into src/Model - Included namespace NightJar\Slug\Model to Slug - Changed arrays() to shortened array syntax "[]" - Update readme.md to reflect changes to SS4 - Changed package type to silverstripe-vendormodule. Composer saves this package into vendor folder - $this->owner->class replaced with $this->owner->getClassName()
